+++ scenes/tree-prusinkiewicz-with-floor.json
@@ -1,0 +1,277 @@
+{
+  "systems": {
+    "tree": {
+      "type": "object",
+      "definition": {
+        "alphabet": [
+          "A(l,w)",
+          "F(l,w)",
+          "[",
+          "]",
+          "+(a)",
+          "/(d)"
+        ],
+        "parameters": {
+          "r1": 0.92,
+          "r2": 0.37,
+          "a1": 0.0,
+          "a2": 60.0,
+          "g1": 180.0,
+          "g2": 0.0,
+          "w0": 2.0,
+          "q": 0.5,
+          "e": 0.0,
+          "min": 0.5
+        },
+        "productions": [
+          "A(l,w): l >= min -> F(l,w)[+(a1)/(g1)A(l*r1,w*Math.pow(q,e))][+(a2)/(g2)A(l*r2,w*Math.pow(1.0-q,e))]"
+        ],
+        "axiom": "A(100,w0)"
+      },
+      "instances": {
+        "a": {
+          "iterations": 10,
+          "parameters": {
+            "r1": 0.75,
+            "r2": 0.77,
+            "a1": 35.0,
+            "a2": -35.0,
+            "g1": 0.0,
+            "g2": 0.0,
+            "w0": 30.0,
+            "q": 0.5,
+            "e": 0.4,
+            "min": 0.0
+          },
+          "materials": [
+            {"color":  [1.0, 1.0, 1.0, 1.0]}
+          ]
+        },
+        "b": {
+          "iterations": 12,
+          "parameters": {
+            "r1": 0.65,
+            "r2": 0.71,
+            "a1": 27.0,
+            "a2": -68.0,
+            "g1": 0.0,
+            "g2": 0.0,
+            "w0": 20.0,
+            "q": 0.53,
+            "e": 0.5,
+            "min": 1.7
+          },
+          "materials": [
+            {"color":  [1.0, 1.0, 1.0, 1.0]}
+          ]
+        },
+        "c": {
+          "iterations": 9,
+          "parameters": {
+            "r1": 0.5,
+            "r2": 0.85,
+            "a1": 25.0,
+            "a2": -15.0,
+            "g1": 180.0,
+            "g2": 0.0,
+            "w0": 20.0,
+            "q": 0.45,
+            "e": 0.5,
+            "min": 0.5
+          },
+          "materials": [
+            {"color":  [1.0, 1.0, 1.0, 1.0]}
+          ]
+        },
+        "d": {
+          "iterations": 10,
+          "parameters": {
+            "r1": 0.6,
+            "r2": 0.85,
+            "a1": 25.0,
+            "a2": -15.0,
+            "g1": 180.0,
+            "g2": 180.0,
+            "w0": 20.0,
+            "q": 0.45,
+            "e": 0.5,
+            "min": 0.0
+          },
+          "materials": [
+            {"color":  [1.0, 1.0, 1.0, 1.0]}
+          ]
+        },
+        "e": {
+          "iterations": 11,
+          "parameters": {
+            "r1": 0.58,
+            "r2": 0.83,
+            "a1": 30.0,
+            "a2": 15.0,
+            "g1": 0.0,
+            "g2": 180.0,
+            "w0": 20.0,
+            "q": 0.4,
+            "e": 0.5,
+            "min": 1.0
+          },
+          "materials": [
+            {"color":  [1.0, 1.0, 1.0, 1.0]}
+          ]
+        },
+        "f": {
+          "iterations": 15,
+          "parameters": {
+            "r1": 0.92,
+            "r2": 0.37,
+            "a1": 0.0,
+            "a2": 60.0,
+            "g1": 180.0,
+            "g2": 0.0,
+            "w0": 2.0,
+            "q": 0.5,
+            "e": 0.0,
+            "min": 0.5
+          },
+          "materials": [
+            {"color":  [1.0, 1.0, 1.0, 1.0]}
+          ]
+        },
+        "g": {
+          "iterations": 10,
+          "parameters": {
+            "r1": 0.8,
+            "r2": 0.8,
+            "a1": 30.0,
+            "a2": -30.0,
+            "g1": 137.0,
+            "g2": 137.0,
+            "w0": 30.0,
+            "q": 0.5,
+            "e": 0.5,
+            "min": 0.0
+          },
+          "materials": [
+            {"color":  [1.0, 1.0, 1.0, 1.0]}
+          ]
+        },
+        "h": {
+          "iterations": 12,
+          "parameters": {
+            "r1": 0.95,
+            "r2": 0.75,
+            "a1": 5.0,
+            "a2": -30.0,
+            "g1": -90.0,
+            "g2": 90.0,
+            "w0": 40.0,
+            "q": 0.6,
+            "e": 0.45,
+            "min": 25.0
+          },
+          "materials": [
+            {"color":  [1.0, 1.0, 1.0, 1.0]}
+          ]
+        },
+        "i": {
+          "iterations": 12,
+          "parameters": {
+            "r1": 0.55,
+            "r2": 0.95,
+            "a1": -5.0,
+            "a2": 30.0,
+            "g1": 137.0,
+            "g2": 137.0,
+            "w0": 5.0,
+            "q": 0.4,
+            "e": 0.0,
+            "min": 5.0
+          },
+          "materials": [
+            {"color":  [1.0, 1.0, 1.0, 1.0]}
+          ]
+        }
+      },
+      "transform": [
+        1, 0, 0, 0,
+        0, 0, 1, 0,
+        0, -1, 0, 0,
+        0, 0, 0, 1
+      ]
+    }
+  },
+  "scene": {
+    "camera": {
+      "eye": [0, 0, -1.5],
+      "lookAt": [0, 0, 0],
+      "up": [0, 1, 0]
+    },
+    "objects": {
+      "Tree": {
+        "type": "lSystem",
+        "system": "tree",
+        "instance": "g",
+        "transform": [
+          1, 0, 0, 0,
+          0, 1, 0, 0,
+          0, 0, 1, 0,
+          0, 0, 0, 1
+        ]
+      },
+      "Floor": {
+        "type": "obj",
+        "obj": "quad.obj",
+        "material": {
+          "color": [1, 1, 1, 1]
+        },
+        "transform": [
+          1, 0, 0, 0,
+          0, 0, 1, 0,
+          0, -1, 0, 0,
+          0, -0.4, 0, 1
+        ]
+      }
+    },
+    "lights": {
+      "ambient": {
+        "color": [
+          0.1,
+          0.1,
+          0.1
+        ]
+      },
+      "pointLights": [
+        {
+          "color": [
+            1.0,
+            1.0,
+            1.0
+          ],
+          "position": [
+            1.0,
+            1.0,
+            0.0
+          ]
+        },
+        {
+          "color": [
+            1.0,
+            1.0,
+            1.0
+          ],
+          "position": [
+            -1.0,
+            1.0,
+            0.0
+          ]
+        }
+      ]
+    }
+  },
+  "resources": {
+    "quad.obj": {
+      "type": "obj",
+      "source": "v -1.000000 1.000000 0.000000\nv 1.000000 1.000000 0.000000\nv -1.000000 -1.000000 0.000000\nv 1.000000 -1.000000 0.000000\n\nvn 0.000000 0.000000 -1.000000\n\nvt 1.000000 1.000000\nvt 0.000000 1.000000\nvt 0.000000 0.000000\nvt 1.000000 0.000000\n\nf 4/4/1 3/3/1 1/2/1\nf 2/1/1 4/4/1 1/2/1"
+    }
+  }
+}
