+++ src/lsystemrenderer/l_system_manager/turtle.rs
@@ -1,36 +1,14 @@
 use crate::framework::context::Gpu;
-use crate::framework::event::lifecycle::Update;
 use crate::framework::geometry::bounds::{Bounds, Bounds3};
 use crate::framework::gpu::buffer::Buffer;
-use crate::framework::input::Input;
 use crate::framework::scene::transform::{Orientation, Transform, Transformable};
-use crate::lindenmayer::LSystem;
 use crate::lsystemrenderer::scene_descriptor::LSystemInstance;
-use crate::lsystemrenderer::turtle::command::TurtleCommand;
-use glam::{Mat4, Quat, Vec3, Vec4};
-use serde::Deserialize;
+use crate::lsystemrenderer::l_system_manager::command::TurtleCommand;
+use glam::{Mat4, Quat, Vec3};
 use std::collections::VecDeque;
 use std::sync::Arc;
 use wgpu::BufferUsages;
-
-#[repr(C)]
-#[derive(Copy, Clone, Debug, Deserialize, bytemuck::Pod, bytemuck::Zeroable)]
-pub struct Material {
-    color: Vec4,
-}
-
-impl Default for Material {
-    fn default() -> Self {
-        Self { color: Vec4::ONE }
-    }
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
-pub struct Instance {
-    matrix: Mat4,
-    material: Material,
-}
+use crate::lsystemrenderer::instancing::{Instance, Material};
 
 pub struct LSystemModel {
     aabb: Bounds3,
@@ -84,14 +62,14 @@ impl TurtleState {
     }
 
     fn make_random_material(&self) -> Material {
-        Material {
-            color: Vec3::new(
+        Material::new(
+            Vec3::new(
                 js_sys::Math::random() as f32,
                 js_sys::Math::random() as f32,
                 js_sys::Math::random() as f32,
             )
             .extend(1.),
-        }
+        )
     }
 
     pub fn get_material(&self) -> Material {
@@ -107,7 +85,7 @@ impl TurtleState {
 }
 
 impl LSystemModel {
-    fn from_turtle_commands(
+    pub fn from_turtle_commands(
         commands: &Vec<TurtleCommand>,
         l_system_transform: Transform,
         initial_material_state: MaterialState,
@@ -131,10 +109,10 @@ impl LSystemModel {
                     let cylinder_transform =
                         Transform::from_scale_rotation(scale_vec, cylinder_base_rotation);
 
-                    cylinder_instances.push(Instance {
-                        matrix: state.transform().as_mat4_with_child(&cylinder_transform), //matrix,
-                        material: state.get_material(),
-                    });
+                    cylinder_instances.push(Instance::new(
+                        state.transform().as_mat4_with_child(&cylinder_transform), //matrix,
+                        state.get_material(),
+                    ));
 
                     state.transform.move_forward(cylinder.length());
 
@@ -201,10 +179,11 @@ impl LSystemModel {
         let model_scale = Mat4::from_scale(Vec3::new(scale_value, scale_value, scale_value));
 
         cylinder_instances.iter_mut().for_each(|c| {
-            c.matrix = model_transform
+            c.set_matrix(model_transform
                 .mul_mat4(&model_scale)
                 .mul_mat4(&model_translation)
-                .mul_mat4(&c.matrix);
+                .mul_mat4(&c.matrix())
+            );
         });
 
         let instances_buffer =
@@ -222,86 +201,5 @@ impl LSystemModel {
 
     pub fn cylinder_instances_buffer(&self) -> &Buffer<Instance> {
         &self.cylinder_instances_buffer
-    }
-}
-
-pub struct LSystemManager {
-    gpu: Arc<Gpu>,
-    max_time_to_iterate: f32,
-    transform: Transform,
-    l_system: LSystem,
-    max_target_iteration: u32,
-    iterations: Vec<LSystemModel>,
-    material_state: MaterialState,
-}
-
-impl LSystemManager {
-    pub fn new(
-        l_system: LSystem,
-        transform: Transform,
-        initial_material_state: Option<MaterialState>,
-        gpu: &Arc<Gpu>,
-    ) -> Self {
-        let mut iterations = Vec::new();
-        let commands: Vec<TurtleCommand> = serde_wasm_bindgen::from_value(l_system.next_raw())
-            .expect("Could not parse turtle commands");
-
-        let material_state = initial_material_state.unwrap_or_default();
-        iterations.push(LSystemModel::from_turtle_commands(
-            &commands,
-            transform,
-            material_state.clone(),
-            gpu,
-        ));
-
-        Self {
-            gpu: gpu.clone(),
-            max_time_to_iterate: 50.,
-            transform,
-            l_system,
-            max_target_iteration: 0,
-            iterations,
-            material_state,
-        }
-    }
-
-    pub fn maybe_increase_max_iteration(&mut self, max_iteration: u32) {
-        self.max_target_iteration = max_iteration.max(self.max_target_iteration);
-    }
-
-    pub fn try_get_iteration(&self, iteration: u32) -> (u32, &LSystemModel) {
-        if self.iterations.len() as u32 > iteration {
-            (iteration, &self.iterations[iteration as usize])
-        } else {
-            let i = self.iterations.len() - 1;
-            (i as u32, &self.iterations[i])
-        }
-    }
-}
-
-impl Update for LSystemManager {
-    fn update(&mut self, input: &Input) {
-        while self.max_target_iteration >= self.iterations.len() as u32 {
-            let commands: Vec<TurtleCommand> =
-                serde_wasm_bindgen::from_value(self.l_system.next_raw())
-                    .expect("Could not parse turtle commands");
-            self.iterations.push(LSystemModel::from_turtle_commands(
-                &commands,
-                self.transform,
-                self.material_state.clone(),
-                &self.gpu,
-            ));
-            if instant::now() as f32 - input.time().now() >= self.max_time_to_iterate {
-                break;
-            }
-        }
-    }
-}
-
-impl Drop for LSystemManager {
-    fn drop(&mut self) {
-        for iteration in self.iterations.iter_mut() {
-            iteration.cylinder_instances_buffer.buffer().destroy();
-        }
     }
 }
